Fixing the scheduler problem
The scheduler used to wake up every 5 milliseconds to check if there is a task that can be assigned to a processor.
This problem was fixed by making the scheduler get notified by the simulator after adding the tasks to the queue.
Now the scheduler no longer has to wake up every 5 milliseconds. Instead it is more efficient now, and wake up when new tasks are added to the queue.